Add files via upload
This is a script for reading fasta file of various close sequences. Filtering them on basis of length. Using Mafft software to align them and give the result in clustal and fasta format. Uses a python script to find the snps and report their positions along with basis in the reference sequnece and the aligned sequence.